pyproject.toml
[tool.poetry]
name = "quasarpy"
version = "0.1.0"
description = "Quasar is python package that can be used for smell detection along with detailed report in various formats such as html, pdf, etc."
authors = ["Khushiyant <khushiyant2002@gmail.com>"]
license = "GNU GENERAL PUBLIC LICENSE"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sphinx = "^7.2.6"
click = "^8.1.7"
pytest = "^8.0.0"
python-dotenv = "^1.0.1"
radon = "^6.0.1"
redis = "^5.0.1"
requests = "^2.31.0"
pygithub = "^2.2.0"
ruff = "^0.2.1"
termcolor = "^2.4.0"
pandas = "^2.2.0"
transformers = "^4.38.1"
torch = "^2.2.1"
pydantic = "^2.6.3"
xgboost = "^2.0.3"
scikit-learn = "^1.4.1.post1"
jinja2 = "^3.1.3"
xhtml2pdf = "^0.2.15"
huggingface-hub = "^0.22.2"
twine = "^5.0.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
extend-include = ["*.ipynb", "__init__.py"]

[radon]
exclude = "tests/*"

